tareas_api/requirements.txt
altgraph==0.17.4
asgiref==3.8.1
contourpy==1.3.2
control==0.10.1
cycler==0.12.1
Django==5.2.3
djangorestframework==3.16.0
fonttools==4.58.0
kiwisolver==1.4.8
matplotlib==3.10.3
numpy==2.2.6
packaging==24.2
pefile==2023.2.7
pillow==11.2.1
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2024.10
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pywin32-ctypes==0.2.3
scipy==1.15.3
setuptools==75.6.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
